backend/requirements.txt
fastapi==0.115.4
uvicorn==0.32.0
pillow==10.4.0
python-multipart==0.0.9
